update of eurparl harvesting code
